services/sentiment-analysis/requirements.txt
feedparser==6.0.12
groq==0.36.0
pandas==2.3.3
tqdm==4.67.1
python-dotenv==1.2.1
beautifulsoup4==4.14.2
lxml==6.0.2
requests==2.32.5
fake-useragent==2.2.0
newspaper3k==0.2.8
lxml_html_clean
streamlit
ollama
yfinance>=0.2.66
plotly
statsmodels>=0.14.0
